Initial nonbuilding comit

+++ src/service/ModelBuilder.Models/Application/PositionAm.cs
@@ -2,7 +2,7 @@ namespace Mb.Models.Application
 {
     public class PositionAm
     {
-        public int PosX { get; set; }
-        public int PosY { get; set; }
+        public double PosX { get; set; }
+        public double PosY { get; set; }
     }
 }
